requirements.txt
blinker==1.7.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.2
distlib==0.3.8
filelock==3.13.1
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
platformdirs==4.2.0
pyasn1==0.5.1
pycparser==2.21
PyJWT==2.8.0
rsa==4.9
virtualenv==20.25.0
Werkzeug==3.0.1
